MGSV thoughts, Language and Scent, the new orientation of the project
MGSV thoughts need organization. The gz.md file can be used for
analyzing the use of particular dates for Extra Ops in Ground Zeroes.
The mgsv.md file has to do with Biblical and Oedipal motifs of
Venom Snake as a king/ruler. I'd like to add thoughts on fathering,
like the way that Venom Snake tells Miller "See? Kid's a natural."

Language and Scent is a reading of Baudelaire.
I think it's somewhere in between phenomenology and semiotics.
The only poem I'm drawing from at the moment is IV. CORRESPONDANCES.
I obviously have more to research.
It was inspired while reading Chapter 2 of Eco's Kant and the
Platypus, but can't seem to remember what. I'll determine this later.

This essay is going to play a role in a project that's been brewing in
my mind for a few months.
I'm concerned with the ontology of AI -- what sort of thing can it be?
Many popular conceptions of AI don't appreciate that it is merely
statistical, merely algorithmic. Deep learning also uses algorithms to
shape the output of neural networks whose structures can be entirely described by
mathematics. In other words, the being of an AI seems to be
arithmetical, and therefore, purely representational. Even the term
quantum computer seems to be cheating a bit. Is the output of a quantum
computer traceable in the same way as that typical computer is?

In any case, I started by saying that AI doesn't have a Kantian imagination.
Because there is no manifold of intuition or sense impressions that a machine
can be aware of, this problem must be overcome by selecting labeled data.
Even unsupervised learning organizes the data along certain axes.
In the famous iris dataset, the length and width of the sepal and petal
are distinguished, for example. Can we say that the machine really
understands these components spatially, or geometrically?
A machine would not be limited to a three-dimensional understanding,
whereas we have a prejudice to limit our ratios to two-term ratios.
It takes creativity to graph two-argument functions, for example.
We can see this in VPD charts. But back to the question of whether the
machine understands measurements geometrically, and whether this is of
fundamental importance to its
We experience space in a non-measured way.
We are often wrong about the lengths of objects in the units of
extension that we've used our whole lives. The intuition of space has no
particular alliance to the arithmetical mode of being. The arithmetical
is merely applicable.
I think the Dedekind essay will necessarily tackle this question,
of whether the geometrical precedes the arithmetical, or vice-versa, in
the form of where the notion of continuity is to be found: in arithmetic
or in geometry. I simply can't make sense of which is prior, or stage a
discussion to determine which must, by the argument, be 'prior' -- and
with respect to inaugurated 'continuity' -- a term which has no exegetic
meaning for me yet.

So AI seems to be a purely arithmetical.
Every sense impression which a human has must have aa
corresponding machinic representation. Monitors must display pixels,
refresh at certain rates...
schemas are not inborn, or latent, but learned. Fast.ai's creator,
Jeremy Howard, in Practical Deep Learning for Coders, showcases a paper
in which a deep learning model shows the features which it has learned
to recognize. This reminded me of schemas, and I think it's worth asking
the question whether this is a fair comparison, and in what way they
differ. For schemas like 'four' and 'dog' (is this a schema?)
So I'm going to have to critically investigate Kantian schemas, and see
whether AI feature-engineering, or a deep-learning model's ability to
detect features on its own, is in any way comparable to the notion of
Kantian schemas. I'll also have to investigate the literature on Kantian
schemas, and see what shortcomings the definition has. I think I recall
some dissatisfactions with its provisional character, but idk.

Is this something that the mind can be? Certainly, the
mind can't be any of machines it produces it produces in the real world,
though the mathematical nature of the AI is identical with some
fundamental part of man. I have to critically read both Dedekind's
"Theory of Numbers", where he introduces the notion of continuity,
and Husserl's "Origin of Geometry" for a future essay, to determine
something on this question.